step2_Interactive_Chat/backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
pydantic-settings==2.1.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
alembic==1.12.1
python-dotenv==1.0.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
langchain-core==0.1.17
numpy==1.26.2
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2 